semantic-segmentation/requirements.txt
altgraph==0.17
click==8.0.1
cycler==0.10.0
dataclasses==0.8
Flask==2.0.1
Flask-Cors==3.0.10
imgviz==1.2.6
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.1.0.25
Pillow==8.3.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PySide2==5.15.2
python-dateutil==2.8.2
PyYAML==5.4.1
shiboken2==5.15.2
six==1.16.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
typing-extensions==3.10.0.1
Werkzeug==2.0.1
zipp==3.5.0
